renames protocol module to robot
