--- opentrons_sdk/protocol/__init__.py
@@ -1,1 +1,0 @@
-from opentrons_sdk.protocol.robot import Robot
