+++ opentrons_sdk/robot/__init__.py
@@ -1,0 +1,1 @@
+from opentrons_sdk.robot.robot import Robot
