+++ tests/opentrons_sdk/integration/test_protocol.py
@@ -1,7 +1,7 @@
 import unittest
 from opentrons_sdk import containers
 
-from opentrons_sdk.protocol import Robot
+from opentrons_sdk.robot import Robot
 from opentrons_sdk.containers.placeable import Container, Deck
 from opentrons_sdk.labware import instruments
